app: rename and split files up + add comments
